+++ src/subscription.jl
@@ -1,9 +1,9 @@
 
-export AbstractSubscription, unsubscribe!
+export Subscription, unsubscribe!
 
 import Base: show
 
-abstract type AbstractSubscription end
+abstract type Subscription end
 
 """
     unsubscribe!(subscription)
@@ -14,7 +14,7 @@ Cancels an associated observable execution and disposes any kind of resources us
 If the input argument to the `unsubscribe!` function is either a tuple or a vector, it will first check that all of the arguments are valid subscription objects 
 and if its true will unsubscribe from each of them individually. 
 
-See also: [`AbstractSubscription`](@ref)
+See also: [`Subscription`](@ref)
 """
 function unsubscribe! end
 
@@ -34,56 +34,3 @@ function unsubscribe! end
 #     end
 #     return nothing
 # end
-
-"""
-    on_unsubscribe!(teardown)
-
-Each valid teardown object with UnsubscribableTeardownLogic trait behavior must implement its own method
-for `on_unsubscribe!()` function which will be invoked when actor decides to `unsubscribe!` from Observable.
-
-See also: [`Teardown`](@ref), [`TeardownLogic`](@ref), [`UnsubscribableTeardownLogic`](@ref)
-"""
-on_unsubscribe!(teardown) = throw(MissingOnUnsubscribeImplementationError(teardown))
-# -------------------------------- #
-# Errors                           #
-# -------------------------------- #
-
-"""
-This error will be thrown if `unsubscribe!` function is called with invalid teardown object.
-
-See also: [`unsubscribe!`](@ref)
-"""
-struct InvalidTeardownLogicTraitUsageError
-    teardown
-end
-
-function Base.show(io::IO, err::InvalidTeardownLogicTraitUsageError)
-    print(io, "Type $(typeof(err.teardown)) has undefined teardown behavior. \nConsider implement as_teardown(::Type{<:$(typeof(err.teardown))}).")
-end
-
-"""
-This error will be thrown if `unsubscribe!` function is called with a tuple with invalid teardown object in it.
-
-See also: [`unsubscribe!`](@ref)
-"""
-struct InvalidMultipleTeardownLogicTraitUsageError 
-    index
-    teardown
-end
-
-function Base.show(io::IO, err::InvalidMultipleTeardownLogicTraitUsageError)
-    print(io, "Check unsubscribe! argument list on index $((err.index)). Type $(typeof(err.teardown)) has undefined teardown behavior. \nConsider implement as_teardown(::Type{<:$(typeof(err.teardown))}).")
-end
-
-"""
-This error will be thrown if Julia cannot find specific method of `on_unsubscribe!()` function for given teardown object.
-
-See also: [`on_unsubscribe!`](@ref)
-"""
-struct MissingOnUnsubscribeImplementationError
-    teardown
-end
-
-function Base.show(io::IO, err::MissingOnUnsubscribeImplementationError)
-    print(io, "You probably forgot to implement on_unsubscribe!(unsubscribable::$(typeof(err.teardown))).")
-end